created tests for calculate js
