+++ package.json
@@ -45,7 +45,9 @@
   "devDependencies": {
     "@babel/core": "^7.20.2",
     "@babel/preset-env": "^7.20.2",
+    "@babel/preset-typescript": "^7.18.6",
     "@testing-library/react": "^13.4.0",
+    "@types/jest": "^29.2.2",
     "babel-jest": "^29.3.1",
     "jest": "^27.5.1"
   }
