--- src/pages/__tests__/calculate.test.js
@@ -1,15 +1,0 @@
-import calculate from '../calculate';
-
-let initialObj = {
-  total: null,
-  next: null,
-  operation: null,
-};
-
-const buttonName = '1';
-
-describe('Test Objects received by calculate', () => {
-  test('Next becomes 1 after sending 1', () => {
-    initialObj = calculate(initialObj, buttonName);
-    expect(initialObj.next).toBe('1');
-  });
